LayoutTests/storage/indexeddb/modern/deletedatabase-request.html

This test calls deleteDatabase on window.indexedDB and verifies that an IDBOpenDBRequest object is returned.
<script>

if (window.testRunner) {
    testRunner.waitUntilDone();
    testRunner.dumpAsText();
}

function done()
{
    alert("Done");
    if (window.testRunner)
        testRunner.notifyDone();
}

var request = window.indexedDB.deleteDatabase("TestDatabase");
alert(request);

done();
</script>
