LayoutTests/storage/indexeddb/modern/deletedatabase-request-expected.txt

ALERT: [object IDBOpenDBRequest]
ALERT: Done
This test calls deleteDatabase on window.indexedDB and verifies that an IDBOpenDBRequest object is returned.
